+++ HaselTweaks/Config/PluginConfig.cs
@@ -122,6 +122,7 @@ public class TweakConfigs
     public AetherCurrentHelperConfiguration AetherCurrentHelper { get; init; } = new();
     public AutoSorterConfiguration AutoSorter { get; init; } = new();
     public BackgroundMusicKeybindConfiguration BackgroundMusicKeybind { get; init; } = new();
+    public BiggerCharacterPreviewsConfiguration BiggerCharacterPreviews { get; init; } = new();
     public CharacterClassSwitcherConfiguration CharacterClassSwitcher { get; init; } = new();
     public CosmicResearchTodoConfiguration CosmicResearchTodo { get; init; } = new();
     public CommandsConfiguration Commands { get; init; } = new();
@@ -154,6 +155,7 @@ public static class PluginConfigExtension
         services.AddSingleton(pluginConfig.Tweaks.AetherCurrentHelper);
         services.AddSingleton(pluginConfig.Tweaks.AutoSorter);
         services.AddSingleton(pluginConfig.Tweaks.BackgroundMusicKeybind);
+        services.AddSingleton(pluginConfig.Tweaks.BiggerCharacterPreviews);
         services.AddSingleton(pluginConfig.Tweaks.CharacterClassSwitcher);
         services.AddSingleton(pluginConfig.Tweaks.CosmicResearchTodo);
         services.AddSingleton(pluginConfig.Tweaks.Commands);